+++ starknet-rpc-test/chain_id.rs
@@ -3,14 +3,12 @@ extern crate starknet_rpc_test;
 use rstest::rstest;
 use starknet_providers::Provider;
 use starknet_rpc_test::constants::SN_GOERLI_CHAIN_ID;
-use starknet_rpc_test::fixtures::madara;
-use starknet_rpc_test::MadaraClient;
+use starknet_rpc_test::fixtures::{madara, ThreadSafeMadaraClient};
 
 #[rstest]
 #[tokio::test]
-async fn returns_hardcoded_chain_id(#[future] madara: MadaraClient) -> Result<(), anyhow::Error> {
-    let madara = madara.await;
-    let rpc = madara.get_starknet_client();
+async fn returns_hardcoded_chain_id(madara: &ThreadSafeMadaraClient) -> Result<(), anyhow::Error> {
+    let rpc = madara.get_starknet_client().await;
 
     assert_eq!(rpc.chain_id().await?, SN_GOERLI_CHAIN_ID);